chore: restructure domain module to have the in-memory implementation of the order book and matching engine self-contained
