+++ ctenex/domain/entities/contract/model.py
