+++ ctenex/domain/entities/trade/model.py
