+++ tests/ctenex/api/v1/test_exchange.py
@@ -4,8 +4,8 @@ from fastapi.encoders import jsonable_encoder
 from fastapi.testclient import TestClient
 
 from ctenex.domain.contracts import ContractCode
-from ctenex.domain.order.model import OrderSide, OrderStatus, OrderType
-from ctenex.domain.order.schemas import OrderAddRequest
+from ctenex.domain.entities.order.model import OrderSide, OrderStatus, OrderType
+from ctenex.domain.entities.order.schemas import OrderAddRequest
 from tests.fixtures import (
     client,  # noqa F401
     limit_buy_order,  # noqa F401
